requirements.txt
# Core
python-dotenv==1.0.1
streamlit==1.33.0

# LangChain ecosystem
langchain==0.2.14
langchain-core==0.2.33
langchain-community==0.2.12
langchain-openai==0.1.22

# Vector / embeddings (optional future use)
chromadb==0.5.5

# OpenAI client
openai==1.40.6
tiktoken==0.7.0

# Google / Gmail API
google-api-python-client==2.137.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Utils
pydantic==2.8.2
requests==2.32.3
httpx==0.27.0